crosstap2: wrapper to build systemtap script on host against given target
crosstap2 is similar to existing crosstap shell script, but unlike crosstap
script that can handle only kernel SystemTap scripts, crosstap2 can also
deal with user land related SystemTap scripts. It discovers and set up
proper --sysroot arguments for stap invocation.

The logic of handling of user-land related cross build arguments is a bit
more complicated. It seemed that to implement it in existing crosstap
script would be a stretch, so crosstap2 is written in python.

Signed-off-by: Victor Kamensky <kamensky@cisco.com>